# assets/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CoinbaseTest": {
				"properties": {
					"testStr": {
						"type": "string"
					}
				},
				"required": [
					"testStr"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CoinbaseWebhookTimeline": {
				"properties": {
					"time": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"time",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CoinbaseWebhook": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"scheduled_for": {
						"type": "string"
					},
					"event": {
						"properties": {
							"data": {
								"properties": {
									"addresses": {
										"properties": {
											"ethereum": {
												"type": "string"
											},
											"bitcoin": {
												"type": "string"
											}
										},
										"required": [
											"ethereum",
											"bitcoin"
										],
										"type": "object"
									},
									"payments": {
										"items": {},
										"type": "array"
									},
									"pricing_type": {
										"type": "string"
									},
									"metadata": {
										"properties": {},
										"type": "object"
									},
									"timeline": {
										"items": {
											"$ref": "#/components/schemas/CoinbaseWebhookTimeline"
										},
										"type": "array"
									},
									"expires_at": {
										"type": "string"
									},
									"created_at": {
										"type": "string"
									},
									"hosted_url": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"code": {
										"type": "string"
									}
								},
								"required": [
									"addresses",
									"payments",
									"pricing_type",
									"metadata",
									"timeline",
									"expires_at",
									"created_at",
									"hosted_url",
									"description",
									"name",
									"code"
								],
								"type": "object"
							},
							"created_at": {
								"type": "string"
							},
							"api_version": {
								"type": "string"
							},
							"type": {
								"type": "string"
							},
							"resource": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"created_at",
							"api_version",
							"type",
							"resource",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"scheduled_for",
					"event"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"Happy",
							"Sad"
						]
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"phoneNumbers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.email-or-name-or-phoneNumbers_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"phoneNumbers"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-name-or-phoneNumbers_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "crypto-pay-kiosk",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Chase Oliphant ",
			"email": "chaseoliphant@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/coinbase/ping": {
			"get": {
				"operationId": "Ping",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CoinbaseTest"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/coinbase/webhook": {
			"post": {
				"operationId": "Webhook",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CoinbaseWebhook"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}